chore: refactor chore components
