+++ src/components/atoms/AnimatedBars.tsx
