+++ src/components/atoms/CustomImage.tsx
