+++ src/components/atoms/InfiniteLoopSlider.tsx
