+++ src/components/sidebar/ProfileHeader.tsx
@@ -1,8 +1,8 @@
 import clsx from 'clsx';
 import Link from 'next/link';
 import { MdVerified as VerifiedIcon } from 'react-icons/md';
-import CustomImage from '../elements/CustomImage';
-import Tooltip from '../elements/Tooltip';
+import CustomImage from '../atoms/CustomImage';
+import Tooltip from '../atoms/Tooltip';
 interface ProfileHeaderProps {
   expandMenu: boolean;
   imageSize: number;
