+++ src/modules/about/components/CareerCard.tsx
@@ -2,8 +2,8 @@ import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
 import { BsBuildings as CompanyIcon } from 'react-icons/bs';
 import { HiChevronRight } from 'react-icons/hi';
-import Card from '@/components/elements/Card';
-import CustomImage from '@/components/elements/CustomImage';
+import Card from '@/components/atoms/Card';
+import CustomImage from '@/components/atoms/CustomImage';
 import { durationBetweenDates, formatDate } from '@/helpers';
 import { CareerProps } from '@/types/careers';
 import { Fragment, useState } from 'react';
