+++ src/modules/dashboard/components/CodingActive/CodingActive.tsx
@@ -5,8 +5,8 @@ import { useEffect, useState } from 'react';
 import { SiWakatime as WakatimeIcon } from 'react-icons/si';
 import useSWR from 'swr';
 
-import SectionHeading from '@/components/elements/SectionHeading';
-import SectionSubHeading from '@/components/elements/SectionSubHeading';
+import SectionHeading from '@/components/atoms/SectionHeading';
+import SectionSubHeading from '@/components/atoms/SectionSubHeading';
 
 import CodingActiveList from './CodingActiveList';
 import Overview from './Overview';
