+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from 'next';
 import { NextSeo } from 'next-seo';
 
-import Container from '@/components/elements/Container';
+import Container from '@/components/atoms/Container';
 import Home from '@/modules/home';
 
 const HomePage: NextPage = () => {
